+++ mailpoet/views/segments/translations.html
@@ -106,7 +106,6 @@
   'emailActionOpensAbsoluteCount': __('number of opens'),
   'emailActionMachineOpensAbsoluteCount': __('number of machine-opens'),
   'emailActionOpens': __('opens'),
-  'emailActionDays': __('days'),
   'emailActionOpensSentence': _x('{condition} {opens} opens', 'The result will be "more than 20 opens"'),
   'emailActionOpensDaysSentence': _x('in the last {days} days', 'The result will be "in the last 5 days"'),
   'moreThan': __('more than'),
